Create Mysql Backup And Put FTP Server
It dumps the mysql database table by table, compresses it in a directory and sends it to a remote ftp server. It also deletes old records in the copies kept in the backup directory. In this way, it does not take up space on the server disk. Even if the FTP connection is lost, the last backup will be on the server where the transaction was made.